Move use() (aka use_item) again
