+++ src/avatar_functions.cpp
@@ -33,6 +33,9 @@ static const itype_id itype_large_repairkit( "large_repairkit" );
 static const itype_id itype_small_repairkit( "small_repairkit" );
 
 static const skill_id skill_weapon( "weapon" );
+
+static const std::string flag_SPLINT( "SPLINT" );
+
 namespace avatar_funcs
 {
 
@@ -539,6 +542,63 @@ void toolmod_add( avatar &you, item_location tool, item_location mod )
     you.assign_activity( activity_id( "ACT_TOOLMOD_ADD" ), 1, -1 );
     you.activity.targets.emplace_back( std::move( tool ) );
     you.activity.targets.emplace_back( std::move( mod ) );
+}
+
+static bool is_pet_food( const item &itm )
+{
+    return itm.type->can_use( "DOGFOOD" ) ||
+           itm.type->can_use( "CATFOOD" ) ||
+           itm.type->can_use( "BIRDFOOD" ) ||
+           itm.type->can_use( "CATTLEFODDER" );
+}
+
+void use_item( avatar &you, item_location loc )
+{
+    item &used = *loc.get_item();
+
+    if( used.is_null() ) {
+        add_msg( m_info, _( "You do not have that item." ) );
+        return;
+    }
+
+    you.last_item = used.typeId();
+
+    if( used.is_tool() ) {
+        if( !used.type->has_use() ) {
+            add_msg( _( "You can't do anything interesting with your %s." ), used.tname() );
+            return;
+        }
+        you.invoke_item( &used, loc.position() );
+
+    } else if( is_pet_food( used ) ) {
+        you.invoke_item( &used, loc.position() );
+
+    } else if( !used.is_container_empty() && is_pet_food( used.get_contained() ) ) {
+        you.unload( loc );
+
+    } else if( !used.is_craft() && ( used.is_medication() || ( !used.type->has_use() &&
+                                     ( used.is_food() ||
+                                       used.get_contained().is_food() ||
+                                       used.get_contained().is_medication() ) ) ) ) {
+        you.consume( loc );
+
+    } else if( used.is_book() ) {
+        you.read( loc );
+    } else if( used.type->has_use() ) {
+        you.invoke_item( &used, loc.position() );
+    } else if( used.has_flag( flag_SPLINT ) ) {
+        ret_val<bool> need_splint = you.can_wear( used );
+        if( need_splint.success() ) {
+            you.wear_item( used );
+            loc.remove_item();
+        } else {
+            add_msg( m_info, need_splint.str() );
+        }
+    } else {
+        add_msg( m_info, _( "You can't do anything interesting with your %s." ),
+                 used.tname() );
+    }
+    you.recalculate_enchantment_cache();
 }
 
 } // namespace avatar_funcs
